+++ src/__generated__/FairHeaderTestsQuery.graphql.ts
@@ -1,0 +1,245 @@
+/* tslint:disable */
+
+import { ConcreteRequest } from "relay-runtime";
+import { FairHeader_fair$ref } from "./FairHeader_fair.graphql";
+export type FairHeaderTestsQueryVariables = {};
+export type FairHeaderTestsQueryResponse = {
+    readonly fair: ({
+        readonly " $fragmentRefs": FairHeader_fair$ref;
+    }) | null;
+};
+export type FairHeaderTestsQuery = {
+    readonly response: FairHeaderTestsQueryResponse;
+    readonly variables: FairHeaderTestsQueryVariables;
+};
+
+
+
+/*
+query FairHeaderTestsQuery {
+  fair(id: "sofa-chicago-2018") {
+    ...FairHeader_fair
+    __id
+  }
+}
+
+fragment FairHeader_fair on Fair {
+  id
+  name
+  image {
+    image_url
+    aspect_ratio
+    url
+  }
+  organizer {
+    profile {
+      icon {
+        id
+        href
+        height
+        width
+        url
+      }
+      name
+      __id
+    }
+  }
+  start_at
+  end_at
+  __id
+}
+*/
+
+const node: ConcreteRequest = (function(){
+var v0 = [
+  {
+    "kind": "Literal",
+    "name": "id",
+    "value": "sofa-chicago-2018",
+    "type": "String!"
+  }
+],
+v1 = {
+  "kind": "ScalarField",
+  "alias": null,
+  "name": "__id",
+  "args": null,
+  "storageKey": null
+},
+v2 = {
+  "kind": "ScalarField",
+  "alias": null,
+  "name": "id",
+  "args": null,
+  "storageKey": null
+},
+v3 = {
+  "kind": "ScalarField",
+  "alias": null,
+  "name": "name",
+  "args": null,
+  "storageKey": null
+},
+v4 = {
+  "kind": "ScalarField",
+  "alias": null,
+  "name": "url",
+  "args": null,
+  "storageKey": null
+};
+return {
+  "kind": "Request",
+  "operationKind": "query",
+  "name": "FairHeaderTestsQuery",
+  "id": "ed40ba0e438ae91a867e8fc168ddfd48",
+  "text": null,
+  "metadata": {},
+  "fragment": {
+    "kind": "Fragment",
+    "name": "FairHeaderTestsQuery",
+    "type": "Query",
+    "metadata": null,
+    "argumentDefinitions": [],
+    "selections": [
+      {
+        "kind": "LinkedField",
+        "alias": null,
+        "name": "fair",
+        "storageKey": "fair(id:\"sofa-chicago-2018\")",
+        "args": v0,
+        "concreteType": "Fair",
+        "plural": false,
+        "selections": [
+          {
+            "kind": "FragmentSpread",
+            "name": "FairHeader_fair",
+            "args": null
+          },
+          v1
+        ]
+      }
+    ]
+  },
+  "operation": {
+    "kind": "Operation",
+    "name": "FairHeaderTestsQuery",
+    "argumentDefinitions": [],
+    "selections": [
+      {
+        "kind": "LinkedField",
+        "alias": null,
+        "name": "fair",
+        "storageKey": "fair(id:\"sofa-chicago-2018\")",
+        "args": v0,
+        "concreteType": "Fair",
+        "plural": false,
+        "selections": [
+          v2,
+          v3,
+          {
+            "kind": "LinkedField",
+            "alias": null,
+            "name": "image",
+            "storageKey": null,
+            "args": null,
+            "concreteType": "Image",
+            "plural": false,
+            "selections": [
+              {
+                "kind": "ScalarField",
+                "alias": null,
+                "name": "image_url",
+                "args": null,
+                "storageKey": null
+              },
+              {
+                "kind": "ScalarField",
+                "alias": null,
+                "name": "aspect_ratio",
+                "args": null,
+                "storageKey": null
+              },
+              v4
+            ]
+          },
+          {
+            "kind": "LinkedField",
+            "alias": null,
+            "name": "organizer",
+            "storageKey": null,
+            "args": null,
+            "concreteType": "organizer",
+            "plural": false,
+            "selections": [
+              {
+                "kind": "LinkedField",
+                "alias": null,
+                "name": "profile",
+                "storageKey": null,
+                "args": null,
+                "concreteType": "Profile",
+                "plural": false,
+                "selections": [
+                  {
+                    "kind": "LinkedField",
+                    "alias": null,
+                    "name": "icon",
+                    "storageKey": null,
+                    "args": null,
+                    "concreteType": "Image",
+                    "plural": false,
+                    "selections": [
+                      v2,
+                      {
+                        "kind": "ScalarField",
+                        "alias": null,
+                        "name": "href",
+                        "args": null,
+                        "storageKey": null
+                      },
+                      {
+                        "kind": "ScalarField",
+                        "alias": null,
+                        "name": "height",
+                        "args": null,
+                        "storageKey": null
+                      },
+                      {
+                        "kind": "ScalarField",
+                        "alias": null,
+                        "name": "width",
+                        "args": null,
+                        "storageKey": null
+                      },
+                      v4
+                    ]
+                  },
+                  v3,
+                  v1
+                ]
+              }
+            ]
+          },
+          {
+            "kind": "ScalarField",
+            "alias": null,
+            "name": "start_at",
+            "args": null,
+            "storageKey": null
+          },
+          {
+            "kind": "ScalarField",
+            "alias": null,
+            "name": "end_at",
+            "args": null,
+            "storageKey": null
+          },
+          v1
+        ]
+      }
+    ]
+  }
+};
+})();
+(node as any).hash = '03c9b173327b501b47b97a6f6817238a';
+export default node;